requirements.txt
torch
torchvision
pytorch-lightning
lightly
pl-bolt
pillow